+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add a `bundled` feature for `gdal-sys` that allows to build and statically link a minimal bundled version of gdal during `cargo build`
+- Add methods ``alternative_name``, ``is_nullable``, ``is_unique``, ``default_value`` to ``Field``.
 
 ## 0.17.1